+++ crates/tasklens-core/src/domain/readiness.rs
@@ -9,7 +9,7 @@
 /// - If no due date, the factor is 1.0.
 /// - Ramp starts at 2 * lead_time remaining.
 /// - Fully ramped at 1 * lead_time remaining.
-pub fn calculate_lead_time_factor(due_date: Option<u64>, lead_time: u64, current_time: u64) -> f64 {
+pub fn calculate_lead_time_factor(due_date: Option<f64>, lead_time: f64, current_time: f64) -> f64 {
     let due_date = match due_date {
         Some(d) => d,
         None => return 1.0,
@@ -21,16 +21,16 @@ pub fn calculate_lead_time_factor(due_date: Option<u64>, lead_time: u64, current
 
     let time_remaining = due_date - current_time;
 
-    if time_remaining > 2 * lead_time {
+    if time_remaining > 2.0 * lead_time {
         return 0.0;
     }
 
-    let raw_factor = ((2 * lead_time) as f64 - time_remaining as f64) / lead_time as f64;
+    let raw_factor = (2.0 * lead_time - time_remaining) / lead_time;
     raw_factor.clamp(0.0, 1.0)
 }
 
 /// Determines if a task is "Ready" to be shown in the UI.
-pub fn is_task_ready(due_date: Option<u64>, lead_time: u64, current_time: u64) -> bool {
+pub fn is_task_ready(due_date: Option<f64>, lead_time: f64, current_time: f64) -> bool {
     match due_date {
         None => true,
         Some(d) => {
@@ -38,7 +38,7 @@ pub fn is_task_ready(due_date: Option<u64>, lead_time: u64, current_time: u64) -
                 true
             } else {
                 let time_remaining = d - current_time;
-                time_remaining <= 2 * lead_time
+                time_remaining <= 2.0 * lead_time
             }
         }
     }
@@ -50,58 +50,58 @@ mod tests {
 
     #[test]
     fn test_calculate_lead_time_factor() {
-        let lead_time = 1000;
-        let due_date = 5000;
+        let lead_time = 1000.0;
+        let due_date = 5000.0;
 
         // Too early (> 2 * lead_time)
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 2000),
+            calculate_lead_time_factor(Some(due_date), lead_time, 2000.0),
             0.0
         );
 
         // At the start of the ramp (exactly 2 * lead_time)
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 3000),
+            calculate_lead_time_factor(Some(due_date), lead_time, 3000.0),
             0.0
         );
 
         // Middle of the ramp
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 3500),
+            calculate_lead_time_factor(Some(due_date), lead_time, 3500.0),
             0.5
         );
 
         // Fully ramped (exactly 1 * lead_time)
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 4000),
+            calculate_lead_time_factor(Some(due_date), lead_time, 4000.0),
             1.0
         );
 
         // Close to due date
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 4500),
+            calculate_lead_time_factor(Some(due_date), lead_time, 4500.0),
             1.0
         );
 
         // Past due date
         assert_eq!(
-            calculate_lead_time_factor(Some(due_date), lead_time, 6000),
+            calculate_lead_time_factor(Some(due_date), lead_time, 6000.0),
             1.0
         );
 
         // No due date
-        assert_eq!(calculate_lead_time_factor(None, lead_time, 6000), 1.0);
+        assert_eq!(calculate_lead_time_factor(None, lead_time, 6000.0), 1.0);
     }
 
     #[test]
     fn test_is_task_ready() {
-        let lead_time = 1000;
-        let due_date = 5000;
+        let lead_time = 1000.0;
+        let due_date = 5000.0;
 
-        assert!(!is_task_ready(Some(due_date), lead_time, 2000));
-        assert!(is_task_ready(Some(due_date), lead_time, 3000));
-        assert!(is_task_ready(Some(due_date), lead_time, 4000));
-        assert!(is_task_ready(Some(due_date), lead_time, 6000));
-        assert!(is_task_ready(None, lead_time, 6000));
+        assert!(!is_task_ready(Some(due_date), lead_time, 2000.0));
+        assert!(is_task_ready(Some(due_date), lead_time, 3000.0));
+        assert!(is_task_ready(Some(due_date), lead_time, 4000.0));
+        assert!(is_task_ready(Some(due_date), lead_time, 6000.0));
+        assert!(is_task_ready(None, lead_time, 6000.0));
     }
 }
